Login and register merge to latest frontend and removed the oldet tjs frontend follder
