+++ frontend/src/assets/slide1.jpg
