+++ frontend/src/pages/Register.css
@@ -8,15 +8,15 @@
 
 body,
 html,
-.login-page {
+.register-page {
   height: 100%;
   width: 100%;
   font-family: 'Poppins', sans-serif;
 }
 
-.login-page {
+.register-page {
   display: flex;
-  height: 100vh;
+  height: 90vh;
   width: 100vw;
 }
 
@@ -42,7 +42,7 @@ html,
   object-fit: cover;
 }
 
-/* --- SLIDER CONTROLS --- */
+/* Slider Controls */
 .slider-controls {
   position: absolute;
   bottom: 20px;
@@ -52,7 +52,6 @@ html,
   pointer-events: none;
 }
 
-/* Positioning Arrows */
 .left-arrow,
 .right-arrow {
   position: absolute;
@@ -82,7 +81,6 @@ html,
   background-color: rgba(0, 0, 0, 0.7);
 }
 
-/* Dots centered below arrows */
 .dots-container {
   position: absolute;
   bottom: 0;
@@ -109,10 +107,9 @@ html,
   background-color: #ffffff;
 }
 
-/* Right Column with form */
+/* Right Column */
 .right-column {
   flex: 1;
-  position: relative;
   display: flex;
   align-items: center;
   justify-content: center;
@@ -127,6 +124,46 @@ html,
   width: 100%;
   max-width: 540px;
   box-shadow: 0 10px 30px rgba(0, 0, 0, 0.08);
+}
+
+.password-wrapper {
+  position: relative;
+}
+
+.password-toggle-icon {
+  position: absolute;
+  right: 15px;
+  top: 40%;
+  transform: translateY(-50%);
+  font-size: 20px;
+  color: #777;
+  cursor: pointer;
+  transition: color 0.3s;
+}
+
+.password-toggle-icon:hover {
+  color: #646cff;
+}
+
+.logo-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-bottom: 30px;
+}
+
+.logo {
+  width: 220px;
+  height: auto;
+  margin-bottom: -30px;
+  margin-top: -40px;
+}
+
+.logo-fallback {
+  display: none;
+  font-size: 24px;
+  font-weight: 600;
+  color: #646cff;
 }
 
 .form-box h2 {
@@ -162,26 +199,6 @@ form input:focus {
   outline: none;
 }
 
-/* Add styling for the password visibility toggle */
-.password-wrapper {
-  position: relative;
-}
-
-.password-toggle-icon {
-  position: absolute;
-  right: 15px;
-  top: 40%;
-  transform: translateY(-50%);
-  font-size: 20px;
-  color: #777;
-  cursor: pointer;
-  transition: color 0.3s;
-}
-
-.password-toggle-icon:hover {
-  color: #646cff;
-}
-
 form button {
   width: 100%;
   padding: 14px;
@@ -215,26 +232,4 @@ form button:hover {
 
 .link-text a:hover {
   text-decoration: underline;
-}
-
-/* Logo Styling */
-.logo-wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 30px;
-}
-
-.logo {
-  width: 220px;
-  height: auto;
-  margin-bottom: -30px;
-  margin-top: -40px;
-}
-
-.logo-fallback {
-  display: none;
-  font-size: 24px;
-  font-weight: 600;
-  color: #646cff;
 }
