--- tjs-frontend/README.md
@@ -1,8 +1,0 @@
-## Teacher' Job Portal
-
-Updated the Complete UI Design for the Job Portal Register and Login pages with additional functionality of password visibility.
-
-install lib - npm i 
-Run - npm run dev 
-
-Install --- npm install react-icons for the icons
