--- tjs-frontend/src/App.jsx
@@ -1,19 +1,0 @@
-import React from "react";
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Login from "./pages/Login";
-import Register from "./pages/Register";
-import Success from "./pages/Success";
-
-function App() {
-  return (
-    <Router>
-      <Routes>
-        <Route path="/" element={<Login />} />
-        <Route path="/register" element={<Register />} />
-        <Route path="/success" element={<Success />} />
-      </Routes>
-    </Router>
-  );
-}
-
-export default App;
